Update v4.3
- Novo sistema mais preciso para contabilização dos pontos minerados;
- Tela Dasboard agora conta com gráfico que exibe o histórico dos pontos minerados;
- Renderização gráfica da interface do software aprimorada para maior qualidade na imagem;
- Adicionados alertas na dashboard;
- Removido botão da tela inicial que sugeria criação de carteira na 4stake (atividades encerradas);
- Adicionado indicador de versão do software próxima a logo do True Mining Desktop;
- Corrigido bug ao tentar retornar ao taskbar icon original;
- Adição de licenças de software para novas bibliotecas utilizadas;
- Outras melhorias e correções de bugs.

+++ Janelas/Popups/ExhangeRates.xaml.cs
@@ -10,7 +10,7 @@ namespace True_Mining_Desktop.Janelas.Popups
     {
         private System.Timers.Timer timerUpdate = new System.Timers.Timer(2000);
 
-        public ExchangeRates(double pointsToCoins)
+        public ExchangeRates(decimal pointsToCoins)
         {
             InitializeComponent();
             new System.Threading.Tasks.Task(() =>
